Minor typing

+++ defaults.go
@@ -7,7 +7,7 @@ const (
 	defaultTunnel          = ""
 	defaultForce           = false
 	defaultDryRun          = false
-	defaultScalingGroup    = uint8(0) // all
+	defaultScalingGroup    = uint(0) // all
 	defaultPrivateRegistry = "arvika.pulcy.com:5000"
 	defaultLogLevel        = "debug"
 )
